Create an api router isolated from heath checks
