+++ lib/cartodb/controllers/factory.js
@@ -55,8 +55,6 @@ const TileTemplateController = require('./template/tile');
 
 const AnalysesController = require('./analyses');
 
-const ServerInfoController = require('./server-info');
-
 module.exports = class ControllersFactory {
     constructor ({ serverOptions, environmentOptions }) {
         const redisOptions = Object.assign({}, environmentOptions.redis, {
@@ -154,8 +152,6 @@ module.exports = class ControllersFactory {
         ['update', 'delete'].forEach(function(eventType) {
             templateMaps.on(eventType, namedMapProviderCache.invalidate.bind(namedMapProviderCache));
         });
-
-        const versions = getAndValidateVersions(serverOptions);
 
         this.mapConfigBasePath = serverOptions.base_url_mapconfig;
         this.templateBasePath = serverOptions.base_url_templated;
@@ -269,13 +265,12 @@ module.exports = class ControllersFactory {
             authApi,
             userLimitsApi
         );
-
-        this.serverInfoController = new ServerInfoController(versions);
     }
 
-    regist (app) {
+    regist (api) {
+        // FIXME: we need a better way to reset cache while running tests
         if (process.env.NODE_ENV === 'test') {
-            app.layergroupAffectedTablesCache = this.layergroupAffectedTablesCache;
+            api.layergroupAffectedTablesCache = this.layergroupAffectedTablesCache;
         }
 
         const mapRouter = router();
@@ -289,7 +284,7 @@ module.exports = class ControllersFactory {
         this.previewTemplateController.register(mapRouter);
         this.analysesController.register(mapRouter);
 
-        app.use(this.mapConfigBasePath, mapRouter);
+        api.use(this.mapConfigBasePath, mapRouter);
 
         const templateRouter = router();
 
@@ -297,13 +292,7 @@ module.exports = class ControllersFactory {
         this.tileTemplateController.register(templateRouter);
         this.adminTemplateController.register(templateRouter);
 
-        app.use(this.templateBasePath, templateRouter);
-
-        const monitorRouter = router();
-
-        this.serverInfoController.register(monitorRouter);
-
-        app.use('/', monitorRouter);
+        api.use(this.templateBasePath, templateRouter);
     }
 };
 
@@ -404,43 +393,4 @@ function createRendererFactory ({ redisPool, serverOptions, environmentOptions }
 
 
     return rendererFactory;
-}
-
-function getAndValidateVersions(options) {
-    // jshint undef:false
-    var warn = console.warn.bind(console);
-    // jshint undef:true
-
-    var packageDefinition = require('../../../package.json');
-
-    var declaredDependencies = packageDefinition.dependencies || {};
-    var installedDependenciesVersions = {
-        camshaft: require('camshaft').version,
-        grainstore: windshaft.grainstore.version(),
-        mapnik: windshaft.mapnik.versions.mapnik,
-        node_mapnik: windshaft.mapnik.version,
-        'turbo-carto': require('turbo-carto').version,
-        windshaft: windshaft.version,
-        windshaft_cartodb: packageDefinition.version
-    };
-
-    var dependenciesToValidate = ['camshaft', 'turbo-carto', 'windshaft'];
-    dependenciesToValidate.forEach(function(depName) {
-        var declaredDependencyVersion = declaredDependencies[depName];
-        var installedDependencyVersion = installedDependenciesVersions[depName];
-        if (declaredDependencyVersion !== installedDependencyVersion) {
-            warn(
-                'Dependency="%s" installed version="%s" does not match declared version="%s". Check your installation.',
-                depName, installedDependencyVersion, declaredDependencyVersion
-            );
-        }
-    });
-
-    // Be nice and warn if configured mapnik version is != installed mapnik version
-    if (windshaft.mapnik.versions.mapnik !== options.grainstore.mapnik_version) {
-        warn('WARNING: detected mapnik version (' + windshaft.mapnik.versions.mapnik + ')' +
-            ' != configured mapnik version (' + options.grainstore.mapnik_version + ')');
-    }
-
-    return installedDependenciesVersions;
 }
