+++ lib/cartodb/server.js
@@ -1,6 +1,7 @@
 const _ = require('underscore');
 const express = require('express');
-const { mapnik } = require('windshaft');
+const windshaft = require('windshaft');
+const { mapnik } = windshaft;
 
 const jsonReplacer = require('./utils/json-replacer');
 
@@ -13,6 +14,7 @@ const cors = require('./middleware/cors');
 const user = require('./middleware/user');
 
 const ControllersFactory = require('./controllers/factory');
+const ServerInfoController = require('./controllers/server-info');
 
 const syntaxError = require('./middleware/syntax-error');
 const errorMiddleware = require('./middleware/error-middleware');
@@ -36,23 +38,37 @@ module.exports = function createServer (serverOptions) {
     app.disable('etag');
     app.set('json replacer', jsonReplacer());
 
-    app.use(logger(serverOptions));
-    app.use(bodyParser.json());
-    app.use(servedByHostHeader());
-    app.use(stats({
+    const api = express.Router();
+
+    api.use(logger(serverOptions));
+    api.use(bodyParser.json());
+    api.use(servedByHostHeader());
+    api.use(stats({
         enabled: serverOptions.useProfiler,
         statsClient: global.statsClient
     }));
-    app.use(lzmaMiddleware());
-    app.use(cors());
-    app.use(user());
+    api.use(lzmaMiddleware());
+    api.use(cors());
+    api.use(user());
 
     const controllers = new ControllersFactory({ serverOptions, environmentOptions: global.environment });
 
-    controllers.regist(app);
+    controllers.regist(api);
 
-    app.use(syntaxError());
-    app.use(errorMiddleware());
+    api.use(syntaxError());
+    api.use(errorMiddleware());
+
+    app.use('/', api);
+
+    const versions = getAndValidateVersions(serverOptions);
+    const serverInfoController = new ServerInfoController(versions);
+
+    serverInfoController.register(app);
+
+    // FIXME: we need a better way to reset cache while running tests
+    if (process.env.NODE_ENV === 'test') {
+        app.layergroupAffectedTablesCache = api.layergroupAffectedTablesCache;
+    }
 
     return app;
 };
@@ -77,4 +93,43 @@ function bootstrapFonts(opts) {
 
 function mapnikVersion(opts) {
     return opts.grainstore.mapnik_version || mapnik.versions.mapnik;
+}
+
+function getAndValidateVersions(options) {
+    // jshint undef:false
+    var warn = console.warn.bind(console);
+    // jshint undef:true
+
+    var packageDefinition = require('../../package.json');
+
+    var declaredDependencies = packageDefinition.dependencies || {};
+    var installedDependenciesVersions = {
+        camshaft: require('camshaft').version,
+        grainstore: windshaft.grainstore.version(),
+        mapnik: windshaft.mapnik.versions.mapnik,
+        node_mapnik: windshaft.mapnik.version,
+        'turbo-carto': require('turbo-carto').version,
+        windshaft: windshaft.version,
+        windshaft_cartodb: packageDefinition.version
+    };
+
+    var dependenciesToValidate = ['camshaft', 'turbo-carto', 'windshaft'];
+    dependenciesToValidate.forEach(function(depName) {
+        var declaredDependencyVersion = declaredDependencies[depName];
+        var installedDependencyVersion = installedDependenciesVersions[depName];
+        if (declaredDependencyVersion !== installedDependencyVersion) {
+            warn(
+                'Dependency="%s" installed version="%s" does not match declared version="%s". Check your installation.',
+                depName, installedDependencyVersion, declaredDependencyVersion
+            );
+        }
+    });
+
+    // Be nice and warn if configured mapnik version is != installed mapnik version
+    if (windshaft.mapnik.versions.mapnik !== options.grainstore.mapnik_version) {
+        warn('WARNING: detected mapnik version (' + windshaft.mapnik.versions.mapnik + ')' +
+            ' != configured mapnik version (' + options.grainstore.mapnik_version + ')');
+    }
+
+    return installedDependenciesVersions;
 }
